+++ src/oneapi/oneapi_backend.h
@@ -1,0 +1,58 @@
+/*******************************************************************************
+* Copyright 2021 Intel Corporation
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*******************************************************************************/
+
+#ifndef __ONEAPI_BACKEND_H_INCLUDED__
+#define __ONEAPI_BACKEND_H_INCLUDED__
+
+#include "daal_sycl.h"
+#ifndef DAAL_SYCL_INTERFACE
+    #include <type_traits>
+    #include <memory>
+static_assert(false, "DAAL_SYCL_INTERFACE not defined")
+#endif
+
+#ifdef _WIN32
+#define ONEAPI_EXPORT __declspec(dllexport)
+#else
+#define ONEAPI_EXPORT
+#endif
+
+class ONEAPI_EXPORT PySyclExecutionContext
+{
+public:
+    // Construct from given device provided as string
+    PySyclExecutionContext(const std::string & dev);
+    ~PySyclExecutionContext();
+
+private:
+    daal::services::SyclExecutionContext * m_ctxt;
+};
+
+template <typename T>
+ONEAPI_EXPORT void* to_device(T * ptr, int * shape);
+
+template <typename T>
+ONEAPI_EXPORT void delete_device_data(void * ptr);
+
+// take a sycl buffer and convert ti oneDAL NT
+template <typename T, bool is_device_data>
+ONEAPI_EXPORT daal::data_management::NumericTablePtr * to_daal_nt(void * ptr, int * shape);
+
+// return a device data from a SyclHomogenNumericTable
+template <typename T>
+ONEAPI_EXPORT void * fromdaalnt(daal::data_management::NumericTablePtr * ptr);
+
+#endif // __ONEAPI_BACKEND_H_INCLUDED__
